Move the code from binary mdbook-i18n-normalize.rs into library code
This refactoring is needed so that fuzzing code can
import and test this logic.
